Refactor Overviews codec
