+++ .changeset/silent-moles-sing.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Refactor Overviews codec
