+++ apps/nextjs-website/src/lib/overviews.ts
@@ -1,4 +1,4 @@
-import { StrapiOverviews } from '@/lib/strapi/overviewsCodec';
+import { StrapiOverviews } from '@/lib/strapi/codecs/overviewsCodec';
 import { OverviewPageProps } from '@/app/[productSlug]/overview/page';
 import { overviews, products } from '@/_contents/products';
