+++ apps/nextjs-website/src/lib/strapi/fetches/fetchOverviews.ts
@@ -1,0 +1,45 @@
+import * as qs from 'qs';
+import { fetchFromStrapi } from '@/lib/strapi/fetchFromStrapi';
+import { OverviewsCodec } from '@/lib/strapi/codecs/overviewsCodec';
+
+const makeStrapiOverviewsPopulate = () =>
+  qs.stringify({
+    populate: {
+      backgroundImage: '*',
+      product: {
+        populate: ['logo'],
+      },
+      relatedLinks: {
+        populate: ['links'],
+      },
+      features: {
+        populate: ['items.icon'],
+      },
+      startInfoSection: {
+        populate: ['bottomLink', 'items.icon'],
+      },
+      tutorialSection: {
+        populate: ['tutorials.image', 'tutorials.product'],
+      },
+      postIntegration: {
+        populate: [
+          'link',
+          'guides.image',
+          'guides.listItems',
+          'guides.mobileImage',
+          'documents.image',
+          'documents.mobileImage',
+          'serviceModels',
+        ],
+      },
+      bannerLinks: {
+        populate: ['icon'],
+      },
+    },
+  });
+
+export const fetchOverviews = fetchFromStrapi(
+  'overviews',
+  makeStrapiOverviewsPopulate(),
+  OverviewsCodec
+);
